scripts/requirements.txt
numpy>=1.26.0
umap-learn>=0.5.5
scipy>=1.12.0
scikit-learn>=1.4.0
